# LeaderboardBackend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "LeaderboardBackend",
    "version": "v1"
  },
  "paths": {
...
    "/api/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Gets a Category by its ID.",
        "operationId": "getCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Gets a Run by its ID.",
        "operationId": "getRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TimedRunViewModel"
                    },
                    {
                      "$ref": "#/components/schemas/ScoredRunViewModel"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "The Run with ID `id` could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}/runs": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Creates a new Run for a Category with ID `id`. This request is restricted to confirmed Users and Administrators.",
        "operationId": "createRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateTimedRunRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CreateScoredRunRequest"
                  }
                ],
                "description": "Request sent when creating a Run. Set `\"time\"` for a timed request\nand `\"score\"` for a scored one."
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/TimedRunViewModel"
                    },
                    {
                      "$ref": "#/components/schemas/ScoredRunViewModel"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The client is not logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "The requesting User is unauthorized to create Runs.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The Category with ID `id` could not be found, or has been deleted. Read `title` for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Gets all Runs submitted by users for a Category. To get only the personal bests of every user instead, call `GetRecordsForCategory`.",
        "operationId": "getRunsForCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of records to return. Fewer records may be returned.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The zero-based index at which to begin selecting records to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StatusFilter"
                }
              ],
              "default": "Published"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunViewModelListView"
                }
              }
            }
          },
          "404": {
            "description": "The Category with ID `id` could not be found, or has been deleted. Read `title` for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}/records": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Gets the records for a category, a.k.a. the personal bests of every user, ranked best-first.",
        "operationId": "getRecordsForCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of records to return. Fewer records may be returned.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The zero-based index at which to begin selecting records to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunViewModelListView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{id}/category": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Gets the category a run belongs to.",
        "operationId": "getRunCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/runs/{id}": {
      "patch": {
        "tags": [
          "Runs"
        ],
        "summary": "Updates a run with the specified new fields. This request is restricted to administrators or users updating their own runs. Note: `runType` cannot be updated. This operation is atomic; if an error occurs, the run will not be updated. All fields of the request body are optional but you must specify at least one.",
        "operationId": "updateRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateTimedRunRequest"
                  },
                  {
                    "$ref": "#/components/schemas/UpdateScoredRunRequest"
                  }
                ],
                "description": "Request sent when updating a run.\nAll fields are optional but you must specify at least one."
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "The user attempted to update another user's run, the user is banned or not yet confirmed, or the user attempted to change the status of a run.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The Run with ID `id` could not be found, or has been deleted. Read `title` for more information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Response can be a `ProblemDetails` for a request that doesn't match the run type of a category, or a `ValidationProblemDetails` otherwise.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Runs"
        ],
        "summary": "Deletes a Run.",
        "operationId": "deleteRun",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^[a-zA-Z0-9-_]{22}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The run does not exist (Not Found) or was already deleted (Already Deleted).\nUse the `title` field of the response to differentiate between the two cases if necessary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CategoryViewModel": {
        "required": [
          "createdAt",
          "deletedAt",
          "id",
          "info",
          "leaderboardId",
          "name",
          "slug",
          "sortDirection",
          "status",
          "type",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier of the `Category`.\n",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "The display name of the `Category`.",
            "example": "Foo Bar Baz%"
          },
          "slug": {
            "type": "string",
            "description": "The URL-scoped unique identifier of the `Category`.\n",
            "example": "foo-bar-baz"
          },
          "info": {
            "type": "string",
            "description": "Information pertaining to the `Category`.",
            "nullable": true,
            "example": "Video proof is required."
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RunType"
              }
            ]
          },
          "sortDirection": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SortDirection"
              }
            ]
          },
          "leaderboardId": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "1984-01-01T00:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "1984-01-01T00:00:00Z"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "1984-01-01T00:00:00Z"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Status"
              }
            ]
          }
        },
        "additionalProperties": false,
        "description": "Represents a `Category` tied to a `Leaderboard`."
      },
...
      "CreateRunRequest": {
        "required": [
          "$type",
          "playedOn"
        ],
        "type": "object",
        "properties": {
          "$type": {
            "type": "string"
          },
          "info": {
            "type": "string"
          },
          "playedOn": {
            "type": "string",
            "description": "The date the `Run` was played on. Must obey the format 'YYYY-MM-DD', with leading zeroes.",
            "format": "date",
            "example": "2000-01-01"
          }
        },
        "additionalProperties": false,
        "description": "Request sent when creating a Run. Set `\"time\"` for a timed request\nand `\"score\"` for a scored one.",
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "Time": "#/components/schemas/CreateTimedRunRequest",
            "Score": "#/components/schemas/CreateScoredRunRequest"
          }
        }
      },
      "CreateScoredRunRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateRunRequest"
          },
          {
            "required": [
              "score"
            ],
            "type": "object",
            "properties": {
              "score": {
                "type": "integer",
                "description": "The score achieved during the run.",
                "format": "int64"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "CreateTimedRunRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CreateRunRequest"
          },
          {
            "required": [
              "time"
            ],
            "type": "object",
            "properties": {
              "time": {
                "type": "string",
                "description": "The duration of the run. Must obey the format 'HH:mm:ss.sss', with leading zeroes.",
                "example": "12:34:56.999"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "RunType": {
        "enum": [
          "Time",
          "Score"
        ],
        "type": "string"
      },
...
      "RunViewModelListView": {
        "required": [
          "data",
          "limitDefault",
          "limitMax",
          "total"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/TimedRunViewModel"
                },
                {
                  "$ref": "#/components/schemas/ScoredRunViewModel"
                }
              ]
            }
          },
          "total": {
            "type": "integer",
            "description": "The total number of records matching the given criteria that\nexist in the database, NOT the total number of records returned.",
            "format": "int64"
          },
          "limitDefault": {
            "type": "integer",
            "description": "The default limit that will be applied for this resource type\nif the client does not specify one in the query string.",
            "format": "int32"
          },
          "limitMax": {
            "type": "integer",
            "description": "The maximum value the client is allowed to specify as a limt for\nendpoints return a paginated list of resources of this type.\nExceeding this value will result in an error.",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ScoredRunViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RunViewModel"
          },
          {
            "required": [
              "score"
            ],
            "type": "object",
            "properties": {
              "score": {
                "type": "integer",
                "description": "The score achieved during the run.",
                "format": "int64"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "SortDirection": {
        "enum": [
          "Ascending",
          "Descending"
        ],
        "type": "string"
      },
...
      "Status": {
        "enum": [
          "Published",
          "Deleted"
        ],
        "type": "string"
      },
      "StatusFilter": {
        "enum": [
          "Published",
          "Deleted",
          "Any"
        ],
        "type": "string"
      },
      "TimedRunViewModel": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RunViewModel"
          },
          {
            "required": [
              "time"
            ],
            "type": "object",
            "properties": {
              "time": {
                "type": "string",
                "description": "The duration of the run.",
                "example": "25:01:01.001"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "UpdateScoredRunRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UpdateRunRequest"
          },
          {
            "type": "object",
            "properties": {
              "score": {
                "type": "integer",
                "format": "int64",
                "nullable": true
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "UpdateTimedRunRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UpdateRunRequest"
          },
          {
            "type": "object",
            "properties": {
              "time": {
                "type": "string",
                "nullable": true,
                "example": "25:01:01.001"
              }
            },
            "additionalProperties": false
          }
        ]
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization using the Bearer scheme",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}